+++ lib/config.php
@@ -1,7 +1,9 @@
 <?php
 /**
- * Roots configuration and constants
+ * Roots configuration
  */
+
+// Enable theme features
 add_theme_support('root-relative-urls');    // Enable relative URLs
 add_theme_support('rewrite-urls');          // Enable URL rewrites
 add_theme_support('h5bp-htaccess');         // Enable HTML5 Boilerplate's .htaccess
@@ -30,15 +32,17 @@ function roots_sidebar_class() {
   echo 'span4';
 }
 
-$get_theme_name = explode('/themes/', get_template_directory());
-define('GOOGLE_ANALYTICS_ID',       ''); // UA-XXXXX-Y
-define('POST_EXCERPT_LENGTH',       40);
-define('WP_BASE',                   wp_base_dir());
-define('THEME_NAME',                next($get_theme_name));
-define('RELATIVE_PLUGIN_PATH',      str_replace(site_url() . '/', '', plugins_url()));
-define('FULL_RELATIVE_PLUGIN_PATH', WP_BASE . '/' . RELATIVE_PLUGIN_PATH);
-define('RELATIVE_CONTENT_PATH',     str_replace(site_url() . '/', '', content_url()));
-define('THEME_PATH',                RELATIVE_CONTENT_PATH . '/themes/' . THEME_NAME);
+// Configuration values
+define('GOOGLE_ANALYTICS_ID', ''); // UA-XXXXX-Y
+define('POST_EXCERPT_LENGTH', 40);
 
-// Set the content width based on the theme's design and stylesheet
+/**
+* $content_width is a global variable used by WordPress for max image upload sizes and media embeds (in pixels)
+*
+* Example: If the content area is 640px wide, set $content_width = 620; so images and videos will not overflow.
+*
+* Default: 940px is the default Bootstrap container width.
+*
+* This is not required or used by Roots.
+*/
 if (!isset($content_width)) { $content_width = 940; }